+++ confik-macros/tests/trybuild.rs
@@ -34,7 +34,7 @@ fn compile_macros() {
     t.compile_fail("tests/trybuild/fail-default-invalid-expr.rs");
     t.compile_fail("tests/trybuild/fail-config-name-value.rs");
     t.compile_fail("tests/trybuild/fail-secret-extra-attr.rs");
-    t.compile_fail("tests/trybuild/fail-derive-literal.rs");
+    t.compile_fail("tests/trybuild/fail-forward-literal.rs");
     t.compile_fail("tests/trybuild/fail-field-from-unknown-type.rs");
     t.compile_fail("tests/trybuild/fail-uncreatable-type.rs");
     t.compile_fail("tests/trybuild/fail-not-a-type.rs");